apply formatter

+++ examples/convert_ani1x.py
@@ -23,10 +23,22 @@ from ANI1x_datasets import dataloader as dl  # ANI1x_datsets from https://github
 
 parser = argparse.ArgumentParser(prog="convert_ani1x_data")
 
-parser.add_argument("-i", "--input_file", action="store", default='./ani1x-release.h5',
-                    help="Location of h5 file for ani1x. Defaults to ./ani1x-release.h5", type=str)
-parser.add_argument("-o", "--output_directory", action="store", type=str, default='.',
-                    help="Directory to put the arrays in. Defaults to cwd")
+parser.add_argument(
+    "-i",
+    "--input_file",
+    action="store",
+    default="./ani1x-release.h5",
+    help="Location of h5 file for ani1x. Defaults to ./ani1x-release.h5",
+    type=str,
+)
+parser.add_argument(
+    "-o",
+    "--output_directory",
+    action="store",
+    type=str,
+    default=".",
+    help="Directory to put the arrays in. Defaults to cwd",
+)
 
 args = parser.parse_args()
 
@@ -37,14 +49,14 @@ path_to_h5file = os.path.abspath(args.input_file)
 
 # List of keys to point to requested data
 # Hardcoded for original ani1x level of theory
-data_keys = ['wb97x_dz.energy', 'wb97x_dz.forces']
+data_keys = ["wb97x_dz.energy", "wb97x_dz.forces"]
 
 max_atoms = 63  # hardcoded for ANI1x
 
 # From CCCBDB, for transforming from raw energies to approximately the atomization energy.
-self_energy_byspecies = {'C': -37.830234, 'H': -0.500608, 'N': -54.568004, 'O': -75.036223}
+self_energy_byspecies = {"C": -37.830234, "H": -0.500608, "N": -54.568004, "O": -75.036223}
 
-number_map = dict(zip([6, 1, 7, 8], 'CHNO'))
+number_map = dict(zip([6, 1, 7, 8], "CHNO"))
 
 number_selfenergy = {k: self_energy_byspecies[v] for k, v in number_map.items()}
 
@@ -53,10 +65,11 @@ hartree_in_kcal = 627.5094740631
 
 ####### Functions
 
+
 def pad_atoms_to(arr, total):
     widths = [[0, 0] for _ in arr.shape]
     widths[1][1] = total - arr.shape[1]
-    return np.pad(arr, widths, constant_values=0, mode='constant')
+    return np.pad(arr, widths, constant_values=0, mode="constant")
 
 
 def repeat_species(arr, n_conf):
@@ -68,31 +81,31 @@ def compute_self_en(conformation_array):
 
 
 ####### Perform the extraction
-sets = {k: [] for k in ['Z', 'R', 'Forces', 'E_total', 'T']}
+sets = {k: [] for k in ["Z", "R", "Forces", "E_total", "T"]}
 
 name_map = {
-    'atomic_numbers': 'Z',
-    'coordinates': 'R',
-    'wb97x_dz.forces': 'Forces',
-    'wb97x_dz.energy': 'E_total',
+    "atomic_numbers": "Z",
+    "coordinates": "R",
+    "wb97x_dz.forces": "Forces",
+    "wb97x_dz.energy": "E_total",
 }
 
-print("Reading arrays...", end='', flush=True)
+print("Reading arrays...", end="", flush=True)
 for data in dl.iter_data_buckets(path_to_h5file, keys=data_keys):
 
-    n_conformations = len(data['coordinates'])
+    n_conformations = len(data["coordinates"])
 
     for k, arr in data.items():
-        if k == 'atomic_numbers':
+        if k == "atomic_numbers":
             self_en = compute_self_en(arr)
             arr = repeat_species(arr, n_conformations)
 
-        if k != 'wb97x_dz.energy':
+        if k != "wb97x_dz.energy":
             arr = pad_atoms_to(arr, max_atoms)
 
         sets[name_map[k]].append(arr)
 
-    sets['T'].append(data['wb97x_dz.energy'] - self_en)
+    sets["T"].append(data["wb97x_dz.energy"] - self_en)
 print("Done!")
 
 print("Post-processing.")
@@ -100,19 +113,19 @@ print("Post-processing.")
 for k in sets:
     sets[k] = np.concatenate(sets[k], axis=0)
 
-for k in ['Forces', 'T']:
+for k in ["Forces", "T"]:
     sets[k] = sets[k] * hartree_in_kcal
 
-sets['Z'] = sets['Z'].astype(int)
-sets['T'] = sets['T'].astype('float32')[:, np.newaxis]
-sets['Grad'] = -sets['Forces']
-del sets['Forces']
-del sets['E_total']
+sets["Z"] = sets["Z"].astype(int)
+sets["T"] = sets["T"].astype("float32")[:, np.newaxis]
+sets["Grad"] = -sets["Forces"]
+del sets["Forces"]
+del sets["E_total"]
 
-print("Saving arrays...", end='', flush=True)
+print("Saving arrays...", end="", flush=True)
 
 for k, v in sets.items():
-    name = f'data-Ani1x_dz{k}.npy'
+    name = f"data-Ani1x_dz{k}.npy"
     name = os.path.join(args.output_directory, name)
     np.save(name, v)
